fixed algorithm to check against each pair of magnitudes specifically. also added other star catalogue folder

+++ star.h
@@ -8,7 +8,7 @@ class Star {
 
  public:
   Point  pt;
-  int    r;
+  double r;
   double x;
   double y;
